Update basic-platform-concepts-38ecf59.md
The link to the pricing of the Extension Suite is not working anymore as the term Extension Suite is also deprecated as far as I know. Therefore I proposed a similar link for information about service to extend SAP solutions. Furthermore I updated the link to the Integration Suite pricing.